Irline Sensor
This repository contains code for a Raspberry Pi Pico microcontroller to interface with infrared (IR) line sensors in order to perform line-following activities in a robotic car. The code shows how to configure and read data from the vehicle's two infrared sensors, which are located on the left and right sides. 
Do note: that this code is purely to test for the working sensor individually, it has not yet been integrated with the other components.